Renamed MetamaskButton to generic WalletButton, and cleaned code
