+++ src/HomePage/Components/Header.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MetamaskButton from "./MetamaskButton";
+import WalletButton from "./WalletButton";
 import "../scss/Header.scss";
 export default function Header(props) {
     const {logo, name} = props;
@@ -10,7 +10,7 @@ export default function Header(props) {
         <img className="logo-container__logo" src={logo} alt="logo"/>
         <div className="logo-container__name"> {name} </div>
       </div>
-      <MetamaskButton />
+      <WalletButton src={process.env.PUBLIC_URL + "/wallet.svg"}/>
     </div>
   );
 }
