+++ src/HomePage/Components/index.js
@@ -1,5 +1,5 @@
 export { default as Header } from "./Header";
 export { default as HomeContent } from "./HomeContent";
 export { default as Layout } from "./Layout";
-export { default as MetamaskButton } from "./MetamaskButton";
+export { default as WalletButton } from "./WalletButton";
 export { default as UnsupportedNetwork } from "./UnsupportedNetwork";
