+++ src/HomePage/scss/Header.scss
@@ -79,8 +79,8 @@
   );
 
   &__image {
-    width: 72px;
-    height: 72px;
+    width: 36px;
+    height: 36px;
   }
   &:hover {
     cursor: pointer;
